services/monitor/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
nats-py==2.7.2
pydantic==2.7.1
httpx==0.27.0
